requirements.txt
requests
numpy
faiss-cpu
sentence-transformers
pandas